[1] Multiset combination as map [2] Mutlset combination for very high frequency

+++ src/test/java/io/github/deepeshpatel/jnumbertools/base/CalculatorTest.java
@@ -233,19 +233,19 @@ public class CalculatorTest {
     void testMultisetCombinationsExact() {
         // Using frequencies {2, 3, 2} (total = 7)
         // k = 0 → only one way (empty selection)
-        assertEquals(1, calculator.multisetCombinationsCount(0, 2, 3, 2));
+        assertEquals(1, calculator.multisetCombinationsCount(0, 2, 3, 2).intValue());
         // k = 2 → expected 6 ways (e.g., [A,A], [A,G], [A,B], [G,G], [G,B], [B,B])
-        assertEquals(6, calculator.multisetCombinationsCount(2, 2, 3, 2));
+        assertEquals(6, calculator.multisetCombinationsCount(2, 2, 3, 2).intValue());
         // k = 3 → expected 8 ways as described in the documentation
-        assertEquals(8, calculator.multisetCombinationsCount(3, 2, 3, 2));
+        assertEquals(8, calculator.multisetCombinationsCount(3, 2, 3, 2).intValue());
         // k = 7 (select all items) → 1 way
-        assertEquals(1, calculator.multisetCombinationsCount(7, 2, 3, 2));
+        assertEquals(1, calculator.multisetCombinationsCount(7, 2, 3, 2).intValue());
         // k = 8 (more than available items) → 0 ways
-        assertEquals(0, calculator.multisetCombinationsCount(8, 2, 3, 2));
+        assertEquals(0, calculator.multisetCombinationsCount(8, 2, 3, 2).intValue());
 
         //Test for large values
-        assertEquals(501501, calculator.multisetCombinationsCount(2000, 1000, 1000, 1000));
-        assertEquals(501501, calculator.multisetCombinationsCount(1000, 1000, 1000, 1000));
+        assertEquals(501501, calculator.multisetCombinationsCount(2000, 1000, 1000, 1000).intValue());
+        assertEquals(501501, calculator.multisetCombinationsCount(1000, 1000, 1000, 1000).intValue());
     }
 
     @Test
